more fixes & refactoring
